backend/requirements.txt
flask
opencv-python
pyzbar
Pillow
requests
flask_cors